rename LoginHeader -> ModalHeader
